+++ application/views/pages/processing/special-form.php
@@ -119,7 +119,7 @@ $(function(){
                                                 .append($('<td>').text(this.total_trip))
                                                 .append($('<td>').attr('id','totalamnt_'+this.id).text(toparseFloat(this.rate * this.total_trip)))
                                             )
-                                            totalamount += parseInt(this.rate) * parseInt(this.total_trip);
+                                            totalamount += parseFloat(this.rate) * parseFloat(this.total_trip);
                                             totaltrips += parseInt(this.total_trip);
 
                 });
